Implement rental management features: add book renting, returning, and rental status retrieval; update models and requests accordingly.

+++ database/migrations/2025_09_17_035628_create_books_table.php
@@ -16,7 +16,7 @@ return new class extends Migration {
             $table->string('title');
             $table->text('description')->nullable();
             $table->string('isbn')->unique()->nullable();
-            $table->integer('pages')->nullable();
+            $table->unsignedBigInteger('pages')->nullable();
             $table->double('price')->nullable();
             $table->string('language', 10)->nullable();
             $table->string('publisher')->nullable();
